backend/requirements.txt
fastapi==0.110.1
uvicorn[standard]==0.29.0
pydantic==1.10.18
pyjwt==2.8.0
python-dotenv==1.0.1
tinydb==4.8.0
typing-extensions==4.12.2
httpx==0.27.0
pytest==8.3.2
pytest-asyncio==0.23.7
mypy==1.10.0
